requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
awscli==1.18.152
axju-jokes==1.0.3
backcall==0.2.0
black==20.8b1
boto3==1.16.17
botocore==1.19.17
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
dataclasses==0.7
decorator==4.4.2
Django==2.0.2
dnspython==1.16.0
docutils==0.15.2
eventlet==0.28.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.1
git-remote-codecommit==1.15.1
greenlet==0.4.17
heroku==0.1.4
idna==2.10
ikp3db==1.4.1
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.1.1
pathspec==0.8.0
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
Pygments==2.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pypng==0.0.20
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.5
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
stevedore==3.2.2
toml==0.10.1
traitlets==4.3.3
tweepy==3.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
